Split linux\windows search classes.
